% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{ent.fun}
\alias{ent.fun}
\title{Primal entropy objective: returns -mean G(w)}
\usage{
ent.fun(w, divergence, r = 1)
}
\arguments{
\item{w}{numeric vector of weights.}

\item{divergence}{character in {"KL","LW","QLS","TS"}.}

\item{r}{numeric tuning for LW/QLS/TS (ignored for KL).}
}
\description{
Primal entropy objective: returns -mean G(w)
}
\keyword{internal}
