% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{lmd.fun}
\alias{lmd.fun}
\title{Dual objective factory for generalized EB}
\usage{
lmd.fun(X, MU, divergence, r = 1)
}
\arguments{
\item{X}{numeric matrix n x p of balancing features (no intercept column).}

\item{MU}{numeric length-p vector of target feature means (no intercept component).}

\item{divergence}{character in {"KL","LW","QLS","TS"}.}

\item{r}{numeric tuning for LW/QLS/TS (ignored for KL).}
}
\value{
A function of \code{lambda} returning the dual objective:
\code{- t(lambda) \%*\% MU + mean( F(X \%*\% lambda) )}, where F depends on the divergence.
}
\description{
Dual objective factory for generalized EB
}
\keyword{internal}
