% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{w.hat.fun}
\alias{w.hat.fun}
\title{Link rho: map linear predictor -> primal weight}
\usage{
w.hat.fun(eta, divergence, r = 1)
}
\arguments{
\item{eta}{numeric vector (linear predictor).}

\item{divergence}{character in {"KL","LW","QLS","TS"}.}

\item{r}{numeric tuning for LW/QLS/TS (ignored for KL).}
}
\description{
Link rho: map linear predictor -> primal weight
}
\keyword{internal}
